added ring size and several products in one object

+++ web/server.py
@@ -10,20 +10,34 @@ import requests
 import telebot
 from telebot.types import InputMediaPhoto
 import config
+import datetime
 
-app = Flask(__name__, template_folder="F:\projects\osty-jewelry-bot\web")
+app = Flask(__name__, template_folder="F:\projects\osty-jewelry-shop\web")
+# app = Flask(__name__, template_folder="/home/ubuntu/jura/web")
 
 
 @app.route("/", methods=["POST", "GET"])
 def home():
     if request.method == "POST":
 
+
         id = request.form["id"]
         amount = request.form["amount"]
         buy_price = request.form["buy_price"]
         sell_price = request.form["sell_price"]
+        type = request.form["type"]
+        name = request.form["name"]
+        ring_size = request.form["ring_size"]
 
-        database.new_record(id, amount, buy_price, sell_price)
+        if database.get_info(id) != []:
+            if type == "ring":
+                for ring in database.get_info(id):
+                    if ring[6] == ring_size:
+                        return render_template("templates/Add.html", alert_ring=True)
+            else:
+                return render_template("templates/Add.html", alert=True)
+
+        database.new_record(id, amount, buy_price, sell_price, type, name, ring_size)
 
         photo_file = request.files["photo"]
 
@@ -42,45 +56,74 @@ def home():
 @app.route("/delete", methods=["GET","POST"], endpoint='func1')
 def delete():
     if request.method == "POST":
-        product_id = int(request.form.get("id"))
 
-        database.delete_record(int(product_id))
-        os.remove(f"static/images/{product_id}.jpg")
+        if request.form["type"] == "ring":
+            product_id = int(request.form.get("id").split(' ')[0])
+            ring_size = int(request.form.get("id").split(' ')[1])
+            database.delete_record(int(product_id), ring_size=ring_size)
+        else:
+            product_id = int(request.form.get("id"))
+            database.delete_record(int(product_id))
+
+        if database.get_info(product_id) == []:
+            os.remove(f"static/images/{product_id}.jpg")
     
     products = list(map(lambda x: x[:-4], os.listdir("static/images")))
-    return render_template("templates/Delete.html", products=products)
+    return render_template("templates/Delete.html", products=database.get_products())
 
 @app.route('/show', methods=["GET","POST"], endpoint="func2")
 def show():
 
-    products = list(map(lambda x: x[:-4], os.listdir("static/images")))
-
     if request.method == "POST":
-        product_id = int(request.form.get("id"))
-        new_info = database.get_info(int(product_id))
-
-        return render_template("templates/Show.html", products=products, new_info=new_info)
     
-    return render_template("templates/Delete.html", products=products)
+        if request.form["type"] == "ring":
+            product_id = int(request.form.get("id").split(' ')[0])
+            ring_size = int(request.form.get("id").split(' ')[1])
+
+            for ring in database.get_info(int(product_id)):
+                if ring[6] == ring_size:
+                    new_info = ring
+        else:
+            product_id = int(request.form.get("id").split)
+            new_info = database.get_info(int(product_id))
+
+
+        return render_template("templates/Show.html", products=database.get_products(), new_info=new_info)
+    
+    return render_template("templates/Delete.html", products=database.get_products())
 
 
 @app.route('/change', methods=["POST", "GET"])
 def change():
     if request.method == "POST":
 
-        product_id = int(request.form['id'])
-        parameter = request.form['parameter']
-        new_value = int(request.form['value'])
+        if request.form["type"] == "ring":
+            product_id = int(request.form.get("id").split(' ')[0])
+            ring_size = int(request.form.get("id").split(' ')[1])
+            parameter = request.form['parameter']
+            if parameter == "name":
+                new_value = request.form['value']
+            else:
+                new_value = int(request.form['value'])
 
-        database.change_info(product_id, parameter, int(new_value))
+            database.change_info(product_id, parameter, new_value, ring_size=ring_size)
+        else:
+            product_id = int(request.form['id'])
+            parameter = request.form['parameter']
+            if parameter == "name":
+                new_value = request.form['value']
+            else:
+                new_value = int(request.form['value'])
+
+
+            database.change_info(product_id, parameter, new_value)
 
     products = list(map(lambda x: x[:-4], os.listdir("static/images")))
-    return render_template("templates/Change.html", products=products)
+    return render_template("templates/Change.html", products=database.get_products())
 
 
 @app.route('/form', methods=["POST", "GET"])
 def send_form():
-    products = list(map(lambda x: x[:-4], os.listdir("static/images")))
 
     if request.method == "POST":
 
@@ -89,17 +132,47 @@ def send_form():
 
         for idx, id in enumerate(id_list):
             if database.amount_info(int(request.form[f"{id}"])) - int(request.form[amount_list[idx]]) < 0:
-                return render_template("templates/SendForm.html", products=products, alert=True)
+                return render_template("templates/SendForm.html", products=database.get_products(), alert=True, product_id=request.form[f"{id}"])
+
 
         add_json(request.form)
 
+
         text = ""
         for i in range(len(id_list)):
-            text += f"""
-Тип: {request.form[f'type{i+1}']}
+
+            product_type = ""
+            if request.form[f'type{i+1}'] == "ring":
+                product_type = "Кільце"
+            
+            if request.form[f'type{i+1}'] == "necklace":
+                product_type = "Намисто"
+
+            if request.form[f'type{i+1}'] == "earrings":
+                product_type = "Cережки"
+            
+            if request.form[f'type{i+1}'] == "bracelette":
+                product_type = "Браслет"
+
+            if request.form[f'type{i+1}'] == "ring":
+
+                for j in database.get_info(request.form[f'id{i+1}']):
+                    if j[6] == int(request.form[f"ringsize{i+1}"]):
+                        text += f"""
+Тип: {product_type}
 ID: {request.form[f'id{i+1}']}
+Ім'я: {j[5]}
 Кількість: {request.form[f'amount{i+1}']}
+Розмір кільця: {request.form[f'ringsize{i+1}']}
 """
+   
+            else:
+                text += f"""
+Тип: {product_type}
+ID: {request.form[f'id{i+1}']}
+Ім'я: {database.get_info(request.form[f'id{i+1}'])[0][5]}
+Кількість: {request.form[f'amount{i+1}']}
+        """
 
         text += f"""
 Ім'я: {request.form['name']}
@@ -120,7 +193,7 @@ Email: {request.form['mail']}
         
         
 
-    return render_template("templates/SendForm.html", products=products)
+    return render_template("templates/SendForm.html", products=database.get_products(), amount=database.get_amount_dict())
 
 @app.route('/orders', methods=["POST", "GET"])
 def manage_orders():
@@ -148,7 +221,12 @@ def confirm_order():
     id_list = list(filter(lambda x: "id" in x, list(orders[order_id].keys())))
 
     for i in range(1, len(id_list)+1):
-        products.append(database.get_info(int(orders[order_id][f"id{i}"])))
+        if orders[order_id][f"type{i}"] == "ring":
+            for ring in database.get_info(int(orders[order_id][f"id{i}"])):
+                if str(ring[6]) == orders[order_id][f"ringsize{i}"]:
+                    products.append(ring)
+        else:
+            products.append(database.get_info(int(orders[order_id][f"id{i}"]))[0])
 
     for i in range(1, len(id_list)+1):
         database.reduce_amount(int(orders[order_id][f"id{i}"]), int(orders[order_id][f"amount{i}"]))
@@ -156,9 +234,11 @@ def confirm_order():
 
     database.add_client([orders[order_id]["name"], orders[order_id]["mail"], orders[order_id]["phone"]])
     
+    for i in products:
+        print(i)
+
     for idx, product in enumerate(products):
-        for _ in range( int(orders[order_id][f"amount{idx+1}"]) ):
-            database.add_profit([product[0], product[2], product[3]])
+        database.add_profit([product[0], product[2], product[3], int(orders[order_id][f"amount{idx+1}"]), str(datetime.date.today()), product[4], product[6]])
     
     # database.add_profit([product[0], product[2], product[3]])
 
@@ -173,6 +253,8 @@ def delete_order():
 
     remove_json(order_id)
     return jsonify(message='Order was deleted')
+
+
 
 @app.route('/edit_order', methods=['POST'])
 def edit_order():
